+++ packages/buttons/src/styled/StyledIconButton.spec.tsx
@@ -10,7 +10,8 @@ import { css } from 'styled-components';
 import { render } from 'garden-test-utils';
 import { COMPONENT_ID, StyledIconButton } from './StyledIconButton';
 import { StyledIcon } from './StyledIcon';
-import { PALETTE_V8 } from '@zendeskgarden/react-theming';
+import { PALETTE } from '@zendeskgarden/react-theming';
+import { rgba } from 'polished';
 
 describe('StyledIconButton', () => {
   it('renders the expected element', () => {
@@ -28,7 +29,7 @@ describe('StyledIconButton', () => {
   it('renders basic color styling', () => {
     const { container } = render(<StyledIconButton isBasic />);
 
-    expect(container.firstChild).toHaveStyleRule('color', PALETTE_V8.grey[600]);
+    expect(container.firstChild).toHaveStyleRule('color', PALETTE.grey[900]);
   });
 
   describe('disabled', () => {
@@ -43,9 +44,13 @@ describe('StyledIconButton', () => {
     it('renders expected primary styling', () => {
       const { container } = render(<StyledIconButton disabled isPrimary />);
 
-      expect(container.firstChild).toHaveStyleRule('background-color', PALETTE_V8.grey[200], {
-        modifier: ':disabled'
-      });
+      expect(container.firstChild).toHaveStyleRule(
+        'background-color',
+        rgba(PALETTE.grey[700], 0.08),
+        {
+          modifier: ':disabled'
+        }
+      );
     });
   });